Reorganized sensor and switch source files
