+++ custom_components/meraki_ha/sensor/device_status.py
@@ -7,7 +7,7 @@ from homeassistant.components.sensor import SensorEntity
 
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 from typing import Dict, Any
-from .const import DOMAIN
+from ..const import DOMAIN
 
 _LOGGER = logging.getLogger(__name__)
